refactor: CarName으로 원시값 포장
