+++ src/main/java/racingcar/domain/RacingCar.java
@@ -1,24 +1,17 @@
 package racingcar.domain;
 
 public class RacingCar {
-    private final String name;
+    private final CarName name;
 
-    private RacingCar(String name) {
+    private RacingCar(CarName name) {
         this.name = name;
     }
 
-    public String getName() {
+    public CarName getName() {
         return name;
     }
 
     public static RacingCar of(String name) {
-        validateLength(name);
-        return new RacingCar(name);
-    }
-
-    private static void validateLength(String name) {
-        if (!(name.length() >= 1 && name.length() <= 5)) {
-            throw new IllegalArgumentException("[ERROR] 차량명은 1글자 이상, 5글자 이하가 되어야합니다");
-        }
+        return new RacingCar(CarName.of(name));
     }
 }
